requirements.txt
flask
Flask
Flask-SQLAlchemy
Flask-Migrate
elastic-apm[flask]
Cython>=0.29.16
kafka-python>=2.0.1
numpy>=1.16.6
pandas>=0.24.2
scipy>=1.4.1
psycopg2-binary
apscheduler
data_linter==0.1.0
pytest==5.4.3
pytest-cov
scikit-learn==0.22.1
